+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.2.0 (unreleased)
 
+- Add a compact nested element style [#1357](https://github.com/AlchemyCMS/alchemy_cms/pull/1357) by [tvdeyen](https://github.com/tvdeyen)
 - Migrating to active_model_serializers ~> 0.10.0 [#1478](https://github.com/AlchemyCMS/alchemy_cms/pull/1478) ([pmashchak](https://github.com/pmashchak))
 
 ## 4.1.0 (2018-09-22)